added api/schemas, utils, config to trim the server file down
